Updates tests and structs.
